Split narrative editor panels into files
